+++ src/ranbval_sdk/__init__.py
@@ -67,6 +67,7 @@ from ranbval_sdk.exceptions import (
     RepoPolicyError,
 )
 from ranbval_sdk.integrations.proxy import aproxy_request, proxy_request
+from ranbval_sdk.remote import fetch_env_set
 from ranbval_sdk.telemetry import (
     aemit_telemetry,
     emit_telemetry,
@@ -76,7 +77,7 @@ from ranbval_sdk.telemetry import (
     uninstall_access_monitor,
 )
 
-__version__ = "3.1.0"
+__version__ = "3.2.0"
 
 __all__ = [
     # Config
@@ -120,6 +121,8 @@ __all__ = [
     # Secure proxy
     "proxy_request",
     "aproxy_request",
+    # Remote config
+    "fetch_env_set",
     # Exceptions
     "RanbvalError",
     "RanbvalDecryptError",